Rename folder to match domain. Cleanup Readme
