+++ README.md
@@ -6,7 +6,7 @@ Support for [Rinnai Control-R Water Heater monitoring and control device](https:
 ![release_date](https://img.shields.io/github/release-date/explosivo22/rinnaicontrolr-ha?style=for-the-badge)
 [![License](https://img.shields.io/github/license/explosivo22/rinnaicontrolr-ha?style=for-the-badge)](https://opensource.org/licenses/Apache-2.0)
 [![hacs_badge](https://img.shields.io/badge/HACS-Custom-orange.svg?style=for-the-badge)](https://github.com/custom-components/hacs)
-
+[![HA integration usage](https://img.shields.io/badge/dynamic/json?color=41BDF5&logo=home-assistant&label=integration%20usage&suffix=%20installs&cacheSeconds=15600&url=https://analytics.home-assistant.io/custom_integrations.json&query=$.rinnai.total)](https://analytics.home-assistant.io/custom_integrations.json)
 <a href="https://www.buymeacoffee.com/Explosivo22" target="_blank"><img src="https://cdn.buymeacoffee.com/buttons/v2/default-blue.png" alt="Buy Me A Coffee" style="height: 60px !important;width: 217px !important;" ></a>
 
 ## KNOWN ISSUES
@@ -56,14 +56,23 @@ Service | Parameters | Description
 The 'master' branch of this custom component is considered unstable, alpha quality and not guaranteed to work.
 Please make sure to use one of the official release branches when installing using HACS, see [what has changed in each version](https://github.com/explosivo22/rinnaicontrolr-ha/releases).
 
-### Install Custom Components
+#### With HACS
 
-1) Make sure that [Home Assistant Community Store (HACS)](https://github.com/custom-components/hacs) is setup.
-2) Go to integrations in HACS
-3) click the 3 dots in the top right corner and choose `custom repositories`
-4) paste the following into the repository input field `https://github.com/explosivo22/rinnaicontrolr-ha` and choose category of `Integration`
-5) click add and restart HA to let the integration load
-6) Recommended to clear the cache and reload first before adding the integration.
-7) Go to settings and choose `Devices & Services`.
-8) Click `Add Integration` and search for `Rinnai Control-R Water Heater`
-9) Configure the integration.
+[![Open your Home Assistant instance and open a repository inside the Home Assistant Community Store.](https://my.home-assistant.io/badges/hacs_repository.svg)](https://my.home-assistant.io/redirect/hacs_repository/?owner=explosivo22&repository=rinnaicontrolr-ha&category=integration)
+
+#### Manual
+1. Copy the `rinnai` directory from `custom_components` in this repository and place inside your Home Assistant's `custom_components` directory.
+2. Restart Home Assistant
+3. Follow the instructions in the `Setup` section
+
+> [!WARNING]
+> If installing manually, in order to be alerted about new releases, you will need to subscribe to releases from this repository.
+
+# Setup
+[![Open your Home Assistant instance and start setting up a new integration.](https://my.home-assistant.io/badges/config_flow_start.svg)](https://my.home-assistant.io/redirect/config_flow_start/?domain=rinnai)
+
+> [!Tip]
+> If you are unable to use the button above, follow the steps below:
+> 1. Navigate to the Home Assistant Integrations page (Settings --> Devices & Services)
+> 2. Click the `+ ADD INTEGRATION` button in the lower right-hand corner
+> 3. Search for `Rinnai Control-R Water Heater`
