+++ custom_components/rinnai/binary_sensor.py
