+++ custom_components/rinnai/entity.py
