--- custom_components/rinnaicontrolr-ha/strings.json
@@ -1,33 +1,0 @@
-{
-    "title": "Rinnai Control-R",
-    "config": {
-        "step": {
-            "user": {
-                "data": {
-                    "host": "[%key:common::config_flow::data::host%]",
-                    "email": "[%key:common::config_flow::data::email%]",
-                    "password": "[%key:common::config_flow::data::password%]"
-                }
-            }
-        },
-        "error": {
-            "cannot_connect": "[%key:common::config_flow::error::cannot_connect%]",
-            "invalid_auth": "[%key:common::config_flow::error::invalid_auth%]",
-            "unknown": "[%key:common::config_flow::error::unknown%]"
-        },
-        "abort": {
-            "already_configured": "[%key:common::config_flow::abort::already_configured_device%]"
-        }
-    },
-    "options": {
-        "step": {
-            "init": {
-              "data": {
-                "units": "Unit of Temperature",
-                "maint_interval_enabled": "Enable scheduled maintenance retrieval (Every 5 min)"
-              },
-                "description": "Configure rinnai options"
-            }
-        }
-    }
-}
